+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/datasources/samples/SamplesDbConnector.groovy
@@ -2,6 +2,7 @@ package life.qbic.portal.sampletracking.datasources.samples
 
 import groovy.util.logging.Log4j2
 import life.qbic.business.DataSourceException
+import life.qbic.business.project.load.LastChangedDateDataSource
 import life.qbic.business.samples.count.CountSamplesDataSource
 import life.qbic.business.samples.download.DownloadSamplesDataSource
 import life.qbic.datamodel.samples.Status
@@ -9,8 +10,8 @@ import life.qbic.portal.sampletracking.datasources.database.ConnectionProvider
 
 import java.sql.Connection
 import java.sql.PreparedStatement
-import java.sql.Statement
 import java.sql.ResultSet
+import java.time.Instant
 
 /**
  * <b>Database connector to the sample tracking database</b>
@@ -23,7 +24,7 @@ import java.sql.ResultSet
  * @since 1.0.0
  */
 @Log4j2
-class SamplesDbConnector implements CountSamplesDataSource, DownloadSamplesDataSource {
+class SamplesDbConnector implements CountSamplesDataSource, DownloadSamplesDataSource, LastChangedDateDataSource {
     private final ConnectionProvider connectionProvider
 
     /**
@@ -93,6 +94,36 @@ class SamplesDbConnector implements CountSamplesDataSource, DownloadSamplesDataS
             }
         }
         return statuses
+    }
+    
+    /**
+     * {@inheritDoc}
+     * @since 1.0.0
+     */
+    @Override
+    Instant getLatestChange(String projectCode) throws DataSourceException {
+      Connection connection = connectionProvider.connect()
+      String latestChangeQuery = "select MAX(UNIX_TIMESTAMP(arrival_time)) from samples_locations where sample_id LIKE ?"
+      String sqlRegex = "${projectCode}%"
+      Instant latest = Instant.MIN
+      connection.withCloseable {
+          PreparedStatement preparedStatement = it.prepareStatement(latestChangeQuery)
+          preparedStatement.setString(1, sqlRegex)
+          ResultSet resultSet = preparedStatement.executeQuery()
+          while (resultSet.next()) {
+            long arrivalTime = resultSet.getLong(1)
+            if(arrivalTime > 0) {
+              try {
+                latest = Instant.ofEpochSecond(arrivalTime)
+              } catch(Exception e) {
+                // The status in the database is invalid. This should never be the case!
+                log.error("Could not parse arrival time $arrivalTime", e)
+                throw new DataSourceException("Retrieval of latest change failed for project $projectCode")
+              }
+            }
+          }
+      }
+      return latest
     }
 
     private class Query {